+++ src/plonetheme/derico/setuphandlers.py
@@ -99,6 +99,7 @@ class HiddenProfiles:
             "plonetheme.derico.upgrades:1004",
             "plonetheme.derico.upgrades:1005",
             "plonetheme.derico.upgrades:1006",
+            "plonetheme.derico.upgrades:1007",
         ]